refactor: 패키지 구조 변경
* auth 패키지 추가 및 관련 파일 이동
