+++ BE/src/main/java/ifive/idrop/auth/filter/AuthController.java
@@ -1,7 +1,7 @@
-package ifive.idrop.controller;
+package ifive.idrop.auth.filter;
 
 import ifive.idrop.dto.request.TokenRefreshRequest;
-import ifive.idrop.jwt.Jwt;
+import ifive.idrop.auth.dto.Jwt;
 import ifive.idrop.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
