+++ BE/src/main/java/ifive/idrop/auth/filter/JwtFilter.java
@@ -1,8 +1,9 @@
-package ifive.idrop.filter;
+package ifive.idrop.auth.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import ifive.idrop.jwt.Jwt;
-import ifive.idrop.jwt.JwtProvider;
+import ifive.idrop.auth.domain.AuthenticateUser;
+import ifive.idrop.auth.dto.Jwt;
+import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.service.UserService;
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
