+++ BE/src/main/java/ifive/idrop/controller/UserController.java
@@ -1,6 +1,6 @@
 package ifive.idrop.controller;
 
-import ifive.idrop.annotation.Login;
+import ifive.idrop.auth.resolver.Login;
 import ifive.idrop.dto.response.BaseResponse;
 import ifive.idrop.dto.response.NameResponse;
 import ifive.idrop.dto.request.SignUpRequest;
