+++ BE/src/main/java/ifive/idrop/service/UserService.java
@@ -9,10 +9,10 @@ import ifive.idrop.entity.Users;
 import ifive.idrop.entity.enums.Role;
 import ifive.idrop.exception.CommonException;
 import ifive.idrop.exception.ErrorCode;
-import ifive.idrop.filter.AuthenticateUser;
-import ifive.idrop.filter.VerifyUserFilter;
-import ifive.idrop.jwt.Jwt;
-import ifive.idrop.jwt.JwtProvider;
+import ifive.idrop.auth.domain.AuthenticateUser;
+import ifive.idrop.auth.filter.VerifyUserFilter;
+import ifive.idrop.auth.dto.Jwt;
+import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
