+++ BE/src/main/java/ifive/idrop/websocket/WebSocketConfig.java
@@ -1,6 +1,6 @@
 package ifive.idrop.websocket;
 
-import ifive.idrop.jwt.JwtProvider;
+import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.repository.UserRepository;
 import ifive.idrop.websocket.direction.NaverDirectionFinder;
 import ifive.idrop.websocket.location.LocationWebSocketHandler;
