+++ BE/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -5,9 +5,9 @@ import ifive.idrop.driver.domain.Driver;
 import ifive.idrop.entity.*;
 import ifive.idrop.exception.CommonException;
 import ifive.idrop.exception.ErrorCode;
-import ifive.idrop.filter.AuthenticateUser;
-import ifive.idrop.filter.VerifyUserFilter;
-import ifive.idrop.jwt.JwtProvider;
+import ifive.idrop.auth.domain.AuthenticateUser;
+import ifive.idrop.auth.filter.VerifyUserFilter;
+import ifive.idrop.auth.utils.JwtProvider;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.pickup.domain.PickUp;
 import ifive.idrop.repository.UserRepository;
